% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpitb.as.data.frame.R
\name{as.data.frame.mpitb_measure}
\alias{as.data.frame.mpitb_measure}
\title{as.data.frame() method for "mpitb_measure" class}
\usage{
\method{as.data.frame}{mpitb_measure}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object of class "mpitb_measure"}

\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{logical}

\item{...}{further arguments passed to or from other methods}
}
\description{
as.data.frame() method for "mpitb_measure" class
}
