% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.mpitb_measure}
\alias{confint.mpitb_measure}
\title{Extract the confident intervals from the calculated measures}
\usage{
\method{confint}{mpitb_measure}(object, parm, level, ...)
}
\arguments{
\item{object}{an object of class `mpitb_measure`.}

\item{parm}{"coefficient"}

\item{level}{numeric}

\item{...}{other arguments}
}
\description{
Extract the confident intervals from the calculated measures
}
