% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{G0.matrix}
\alias{G0.matrix}
\title{Obtain the deprivations matrix (G0) from the survey data.}
\usage{
G0.matrix(data, indicators)
}
\arguments{
\item{data}{a `survey.design` object containing the complex household survey design}

\item{indicators}{a character vector with the names of the indicators.}
}
\value{
a numeric matrix with binary entries representing the deprivations matrix
}
\description{
Obtain the deprivations matrix (G0) from the survey data.
}
\examples{
library(mpitb)
data <- survey::svydesign(id=~PSU, weights = ~Weight, strata = ~Strata,
        data = swz_mics14)
indicators <- c("Water","Assets","School","Nutrition")

deprivations.matrix <- G0.matrix(data, indicators)
}
