% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpitb.A.R, R/mpitb.M0.R
\name{mpitb.A}
\alias{mpitb.A}
\alias{mpitb.A.mpitb_set}
\title{Intensity (A)}
\usage{
mpitb.A(object, ...)

\method{mpitb.A}{mpitb_set}(object, ...)

mpitb.A(object, ...)

\method{mpitb.A}{mpitb_set}(object, ...)
}
\arguments{
\item{object}{`mpitb_set` object}

\item{...}{other arguments}
}
\value{
`mpitb_A` and `mpitb_measure` class

`mpitb_A` and `mpitb_measure` class
}
\description{
Calculates the Intensity of poverty or the average deprivation score of the multidimensionally poor people

Calculates the Intensity of poverty or the average deprivation score of the multidimensionally poor people
}
