% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpitb.change.abs.R
\name{mpitb.change_abs}
\alias{mpitb.change_abs}
\alias{mpitb.change_abs.default}
\title{Absolute rate of change}
\usage{
mpitb.change_abs(period_t1, period_t2, ...)

\method{mpitb.change_abs}{default}(period_t1, period_t2, ...)
}
\arguments{
\item{period_t1}{a \code{mpitb_measure} class object.}

\item{period_t2}{a \code{mpitb_measure} class object. Should be the same AF measure class as `period_t1` with the same parameters (cutoffs, subgroups, etc.).}

\item{...}{other arguments}
}
\value{
\code{mpitb_change} and \code{mpitb_measure} class object
}
\description{
Calculate the absolute rate of change between two AF measures (e.g. \eqn{M_0}) which is simply the difference between the final \eqn{t^2} and the initial period \eqn{t^2}
\deqn{\Delta(M_0)= M_0^{t^2} - M_0^{t^1}}
}
\references{
\emph{Alkire, S., Roche, J. M., & Vaz, A. (2017). Changes over time in multidimensional poverty: Methodology and results for 34 countries. World Development, 94, 232-249}. 10.1016/j.worlddev.2017.01.011
}
