% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpitb.change.rel.R
\name{mpitb.change_rel}
\alias{mpitb.change_rel}
\alias{mpitb.change_rel.default}
\title{Relative rate of change}
\usage{
mpitb.change_rel(period_t1, period_t2, ...)

\method{mpitb.change_rel}{default}(period_t0, period_t1, ...)
}
\arguments{
\item{period_t1}{an \code{mpitb_measure} class object.}

\item{period_t2}{a \code{mpitb_measure} class object. Should be the same AF measure class as `period_t1` with the same parameters (cutoffs, subgroups, etc.).}
}
\value{
\code{mpitb_change} and \code{mpitb_measure} class object
}
\description{
Calculate the relative rate of change between two AF measures (e.g. \eqn{M_0}) which is the difference in poverty as a percentage of the initial poverty level \eqn{t^1}.
\deqn{\delta(M_0)= \frac{M_0^{t^2} - M_0^{t^1}}{M_0^{t^1}} \times 100 }
}
\references{
\emph{Alkire, S., Roche, J. M., & Vaz, A. (2017). Changes over time in multidimensional poverty: Methodology and results for 34 countries. World Development, 94, 232-249}. 10.1016/j.worlddev.2017.01.011
}
