% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpitb.change_rel_ann.R
\name{mpitb.change_rel_ann}
\alias{mpitb.change_rel_ann}
\alias{mpitb.change_rel_ann.default}
\title{Annualized relative rate of change}
\usage{
mpitb.change_rel_ann(period_t1, period_t2, ...)

\method{mpitb.change_rel_ann}{default}(period_t1, period_t2)
}
\arguments{
\item{period_t1}{a \code{mpitb_measure} class object.}

\item{period_t2}{a \code{mpitb_measure} class object. Should be the same AF measure class as `period_t1` with the same parameters (cutoffs, subgroups, etc.).}

\item{...}{other arguments}
}
\value{
\code{mpitb_change} and \code{mpitb_measure} class object
}
\description{
Calculate the annualized relative rate of an AF measure (e.g. \eqn{M_0}) which is the compound rate of reduction in per year between the initial and the final periods.
\deqn{\bar{\delta}(M_0)= \left[ \left( \frac{M_0^{t^2}}{M_0^{t^1}} \right)^{\frac{1}{t^2 - t^1}} - 1 \right] \times 100 }
}
\references{
\emph{Alkire, S., Roche, J. M., & Vaz, A. (2017). Changes over time in multidimensional poverty: Methodology and results for 34 countries. World Development, 94, 232-249}. 10.1016/j.worlddev.2017.01.011
}
