% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpitb.headcounts.R
\name{mpitb.headcounts}
\alias{mpitb.headcounts}
\alias{mpitb.headcounts.mpitb_set}
\title{Censored Headcount Ratios}
\usage{
mpitb.headcounts(object, ...)

\method{mpitb.headcounts}{mpitb_set}(object, censored = FALSE, ...)
}
\arguments{
\item{object}{`mpitb_set` object}

\item{...}{other arguments}

\item{censored}{logical. If TRUE, it calculates the censored headcounts ratios. It is set FALSE by default.}
}
\value{
`mpitb_headcounts` and `mpitb_measure` class
}
\description{
Calculate the censored headcount ratios for each indicator or the proportion of people who are multidimensionally poor and deprived in each indicator
}
\examples{
library(mpitb)
data <- survey::svydesign(id=~PSU, weights = ~Weight, strata = ~Strata,
        data = swz_mics14)
indicators <- c("Water","Assets","School","Nutrition")
weights <- c(1/6,1/6,1/3,1/3)
cutoff <- c(25,50)
over <- c("Region","Area")

set <- mpitb.set(data, indicators, cutoff, weights, over,
      name = "Example", desc = "SWZ MICS survey 2014")

h_j <- mpitb.headcounts(set)
as.data.frame(h_j)
}
