% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpitb.M0.R
\name{mpitb.M0}
\alias{mpitb.M0}
\alias{mpitb.M0.mpitb_set}
\title{Adjusted Headcount Ratio (\eqn{M_0})}
\usage{
mpitb.M0(object, ...)

\method{mpitb.M0}{mpitb_set}(object, ...)
}
\arguments{
\item{object}{a \code{mpitb_set} class object.}

\item{...}{other arguments}
}
\value{
\code{mpitb_M0} and \code{mpitb_measure} class objects
}
\description{
Calculate the Adjusted Headcount Ratio (\eqn{M_0})
\deqn{M_0 = \frac{1}{n} \sum_{i=1}^n c_i(k) = \frac{q}{n} \times \frac{1}{q} \sum_{i=1}^n c_i(k) = H \times A} where  \eqn{c_i(k) = \sum_{j=1}^d w_j d_{ij}(k)} is the censored deprivation score
for each observation, \eqn{n} is the total number of observations, and \eqn{q} is the number of people identified as multidimensional poor.
}
\examples{
library(mpitb)
data <- survey::svydesign(id=~PSU, weights = ~Weight, strata = ~Strata,
        data = swz_mics14)
indicators <- c("Water","Assets","School","Nutrition")
weights <- c(1/6,1/6,1/3,1/3)
cutoff <- c(25,50)
subgroup <- c("Region","Area")

set <- mpitb.set(data, indicators, weights, cutoff, subgroup,
      name = "Example", desc = "SWZ MICS survey 2014")

M0 <- mpitb.M0(set)

## to observe the results in a data.frame format
as.data.frame(M0)
}
\references{
\emph{Alkire, S., Foster, J. E., Seth, S., Santos, M. E., Roche, J., & Ballon, P. (2015). Multidimensional poverty measurement and analysis. Oxford University Press.}
}
