% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpitb.measure.R
\name{mpitb.measure}
\alias{mpitb.measure}
\alias{mpitb.measure.mpitb_set}
\title{Alkire-Foster measure}
\usage{
mpitb.measure(object, ...)

\method{mpitb.measure}{mpitb_set}(
  object,
  measure = c("M0", "H", "A", "censored.headcounts", "headcounts"),
  ...
)
}
\arguments{
\item{object}{`mpitb_set` object}

\item{...}{other arguments}

\item{measure}{desired AF measure. The options are "M0", "H", "A", "censored.headcounts", "headcounts".}
}
\value{
a `mpitb_measure` class object
}
\description{
Calculate any available AF measure (Adjusted Headcount Ratio - M0 -, Incidence - H -, Intensity - A - and censored and uncensored indicators headcount ratios )
}
\examples{
library(mpitb)

data_syn <- syn_cdta
data_syn1 <- subset(data_syn, t == 1)
data_syn1 <- survey::svydesign(id=~psu, weights = ~weight , strata = ~stratum, data = data_syn1)

indicators <- c("d_nutr","d_cm","d_satt","d_educ", "d_elct","d_sani","d_wtr","d_hsg","d_ckfl","d_asst")
weights <- c(1/6,1/6,1/6,1/6, 1/18,1/18,1/18,1/18,1/18,1/18)
cutoff <- c(33,50)
over <- c("region","area")
set <- mpitb.set(data_syn1, indicators, weights, cutoff, subgroup = over, name = "syn_data", desc = "This is an example taken from the synthetic cleaned data from mpitb STATA package")

# Asumme we want to calculate M0 measure

M0<- mpitb.measure(set, measure = "M0")
}
