% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpitb.set.R
\name{mpitb.set}
\alias{mpitb.set}
\title{Set the specification of the project}
\usage{
mpitb.set(
  data,
  indicators,
  K,
  weights,
  over = NULL,
  year,
  name = NULL,
  desc = NULL,
  ...
)
}
\arguments{
\item{data}{`survey.design` object}

\item{indicators}{Vector or a List.}

\item{K}{Numerical. Vector of poverty cut-offs. Can be values between 1 and 100.}

\item{weights}{Weights of each indicator.}

\item{over}{Character.}

\item{year}{Character.}

\item{name}{Character.}

\item{desc}{Character.}

\item{...}{other arguments}
}
\value{
`mpitb_set` object
}
\description{
Set the specification of the project
}
\examples{

library(mpitb)
data <- survey::svydesign(id=~PSU, weights = ~Weight, strata = ~Strata,
        data = swz_mics14)
indicators <- c("Water","Assets","School","Nutrition")
weights <- c(1/6,1/6,1/3,1/3)
cutoff <- c(25,50)
over <- c("Region","Area")

set <- mpitb.set(data, indicators, cutoff, weights, over,
      name = "Example", desc = "SWZ MICS survey 2014")
}
