% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mpitb_measure.R
\name{print.mpitb_measure}
\alias{print.mpitb_measure}
\title{print() method for "mpitb_measure" class}
\usage{
\method{print}{mpitb_measure}(x, ...)
}
\arguments{
\item{x}{an object of class "mpitb_measure"}

\item{...}{other arguments passed to or from other methods}
}
\description{
print() method for "mpitb_measure" class
}
