% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mpitb_set}
\alias{print.mpitb_set}
\title{print() method for "mpitb_set" class}
\usage{
\method{print}{mpitb_set}(object, ...)
}
\arguments{
\item{object}{An object of class "mpitb_set"}

\item{...}{Other arguments passed to or from other methods}
}
\description{
print() method for "mpitb_set" class
}
