% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{swz_mics14}
\alias{swz_mics14}
\title{eSwatini MICS 2014 merged}
\format{
An object of class \code{data.frame} with 20739 rows and 11 columns.
}
\usage{
swz_mics14
}
\description{
eSwatini MICS 2014 merged
}
\references{
\url{https://mics.unicef.org/surveys}
}
\author{
OPHI Summer School 2022
}
\keyword{datasets}
