% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.mpitb_measure}
\alias{coef.mpitb_measure}
\title{Extract the coefficients from the calculated measures}
\usage{
\method{coef}{mpitb_measure}(object, k = NULL, over = NULL, level = NULL, ...)
}
\arguments{
\item{object}{an object of class `mpitb_measure`.}

\item{k}{numeric value vector of the poverty cut-offs (k) for which the coefficients are desired.}

\item{over}{character}

\item{level}{character}

\item{...}{other arguments}
}
\description{
Extract the coefficients from the calculated measures
}
