% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpitb.coef.R
\name{coef.mpitb_measure}
\alias{coef.mpitb_measure}
\title{Extract the coefficients from the calculated measures}
\usage{
\method{coef}{mpitb_measure}(object, k = NULL, over = NULL, level = NULL, ...)
}
\arguments{
\item{object}{an object of class `mpitb_measure`.}

\item{k}{numeric value vector of the poverty cut-offs (k) for which the coefficients are desired.}

\item{over}{character}

\item{level}{character}

\item{...}{other arguments}
}
\description{
Extract the coefficients from the calculated measures
}
\examples{

library(mpitb)
data <- survey::svydesign(id=~PSU, weights = ~Weight, strata = ~Strata,
        data = swz_mics14)
indicators <- c("Water","Assets","School","Nutrition")
weights <- c(1/6,1/6,1/3,1/3)
cutoff <- c(25,50)
over <- c("Region","Area")

set <- mpitb.set(data, indicators, weights, cutoff, over,
      name = "Example", description = "SWZ MICS survey 2014")

M0 <- mpitb.M0(set)
coef(M0)

}
