% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.mpitb_measure.R
\name{confint.mpitb_measure}
\alias{confint.mpitb_measure}
\title{Extract the confident intervals from the calculated measures}
\usage{
\method{confint}{mpitb_measure}(object, parm = "coefficient", level = 0.95, ...)
}
\arguments{
\item{object}{an object of class `mpitb_measure`.}

\item{parm}{"coefficient". Confidence intervals are only available for AF measure point estimates.}

\item{level}{the confidence level required.}

\item{...}{additional argument(s) for methods.}
}
\description{
Extract the confident intervals from the calculated measures
}
\examples{

library(mpitb)
data <- survey::svydesign(id=~PSU, weights = ~Weight, strata = ~Strata,
        data = swz_mics14)
indicators <- c("Water","Assets","School","Nutrition")
weights <- c(1/6,1/6,1/3,1/3)
cutoff <- c(25,50)
over <- c("Region","Area")

set <- mpitb.set(data, indicators, weights, cutoff, over,
      name = "Example", desc = "SWZ MICS survey 2014")

M0 <- mpitb.M0(set)

confint(M0, parm = "coefficient", level = 0.95)
}
