% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.mpitb_measure}
\alias{confint.mpitb_measure}
\title{Extract the confident intervals from the calculated measures}
\usage{
\method{confint}{mpitb_measure}(object, parm, level, ...)
}
\arguments{
\item{object}{an object of class `mpitb_measure`.}

\item{parm}{"coefficient"}

\item{level}{numeric}

\item{...}{other arguments}
}
\description{
Extract the confident intervals from the calculated measures
}
\examples{

library(mpitb)
data <- survey::svydesign(id=~PSU, weights = ~Weight, strata = ~Strata,
        data = swz_mics14)
indicators <- c("Water","Assets","School","Nutrition")
weights <- c(1/6,1/6,1/3,1/3)
cutoff <- c(25,50)
over <- c("Region","Area")

set <- mpitb.set(data, indicators, cutoff, weights, over,
      name = "Example", desc = "SWZ MICS survey 2014")

M0 <- mpitb.M0(set)
as.data.frame(M0)

confint(M0, parm = "coefficient", level = 0.95)
}
