% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{deprivations.score}
\alias{deprivations.score}
\title{Calculate deprivations score from the deprivation matrix}
\usage{
deprivations.score(x)
}
\arguments{
\item{x}{Numeric Matrix}
}
\value{
A numeric vector
}
\description{
Calculate deprivations score from the deprivation matrix
}
