% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpitb.est.R
\name{mpitb.est}
\alias{mpitb.est}
\alias{mpitb.est.mpitb_set}
\title{Estimation of Alkire-Foster measures}
\usage{
mpitb.est(object, measures, ...)

\method{mpitb.est}{mpitb_set}(object, measures = c("M0", "H", "A", "Headcounts"), ...)
}
\arguments{
\item{object}{a `mpitb_set` class object}

\item{measures}{Character vector with the desired AF measures. By default: c("M0", "A", "H", "Headcounts")}

\item{...}{other arguments}
}
\description{
Estimation of Alkire-Foster measures
}
