% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpitb.H.R
\name{mpitb.H}
\alias{mpitb.H}
\alias{mpitb.H.mpitb_set}
\title{Incidence (H)}
\usage{
mpitb.H(object, ...)

\method{mpitb.H}{mpitb_set}(object, ...)
}
\arguments{
\item{object}{`mpitb_set` object}

\item{...}{other arguments}
}
\value{
`mpitb_H` and `mpitb_measure` class
}
\description{
Calculate the Incidence of poverty or the proportion of multidimensionally poor people.
}
