% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpi.R
\name{mpitb.headcounts}
\alias{mpitb.headcounts}
\alias{mpitb.headcounts.mpitb_set}
\title{Censored Headcount Ratios}
\usage{
mpitb.headcounts(object, ...)

\method{mpitb.headcounts}{mpitb_set}(object, ...)
}
\arguments{
\item{object}{`mpitb_set` object}

\item{...}{other arguments}
}
\value{
`mpitb_headcounts` and `mpitb_measure` class
}
\description{
Calculate the censored headcount ratios for each indicator or the proportion of people who are multidimensionally poor and deprived in each indicator
}
