% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpitb.M0.R, R/mpitb.est.R, R/mpitb.headcounts.R
\name{mpitb.headcounts}
\alias{mpitb.headcounts}
\alias{mpitb.headcounts.mpitb_set}
\alias{mpitb.est}
\alias{mpitb.est.mpitb_set}
\title{Censored Headcount Ratios}
\usage{
mpitb.headcounts(object, ...)

\method{mpitb.headcounts}{mpitb_set}(object, ...)

mpitb.est(object, measures, ...)

\method{mpitb.est}{mpitb_set}(object, measures = c("M0", "H", "A", "Headcounts"), ...)

mpitb.headcounts(object, ...)

\method{mpitb.headcounts}{mpitb_set}(object, ...)
}
\arguments{
\item{object}{`mpitb_set` object}

\item{...}{other arguments}

\item{measures}{Character vector with the desired AF measures. By default: c("M0", "A", "H", "Headcounts")}
}
\value{
`mpitb_headcounts` and `mpitb_measure` class

`mpitb_headcounts` and `mpitb_measure` class
}
\description{
Calculate the censored headcount ratios for each indicator or the proportion of people who are multidimensionally poor and deprived in each indicator

Calculate the censored headcount ratios for each indicator or the proportion of people who are multidimensionally poor and deprived in each indicator
}
