% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpi.R
\name{mpitb.M0}
\alias{mpitb.M0}
\alias{mpitb.M0.mpitb_set}
\title{Adjusted Headcount Ratio (M0)}
\usage{
mpitb.M0(object, ...)

\method{mpitb.M0}{mpitb_set}(object, ...)
}
\arguments{
\item{object}{`mpitb_set` object}

\item{...}{other arguments}
}
\value{
`mpitb_M0` and `mpitb_measure` class
}
\description{
Calculate the Adjusted Headcount Ratio (M0 Alkire-Foster class measure)
}
