% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpitb.M0.R
\name{mpitb.M0}
\alias{mpitb.M0}
\alias{mpitb.M0.mpitb_set}
\title{Adjusted Headcount Ratio (M0)}
\usage{
mpitb.M0(object, ...)

\method{mpitb.M0}{mpitb_set}(object, ...)
}
\arguments{
\item{object}{`mpitb_set` object}

\item{...}{other arguments}
}
\value{
`mpitb_M0` and `mpitb_measure` class
}
\description{
Calculate the Adjusted Headcount Ratio (M0 Alkire-Foster class measure)
}
\examples{
library(mpitb)
data <- survey::svydesign(id=~PSU, weights = ~Weight, strata = ~Strata,
        data = swz_mics14)
indicators <- c("Water","Assets","School","Nutrition")
weights <- c(1/6,1/6,1/3,1/3)
cutoff <- c(25,50)
subgroup <- c("Region","Area")

set <- mpitb.set(data, indicators, weights, cutoff, subgroup,
      name = "Example", desc = "SWZ MICS survey 2014")

M0 <- mpitb.M0(set)
as.data.frame(M0)
}
