% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{mpitb.set}
\alias{mpitb.set}
\title{Set the specification of the project}
\usage{
mpitb.set(
  data,
  indicators,
  K,
  weights,
  over = NULL,
  name = NULL,
  desc = NULL,
  ...
)
}
\arguments{
\item{data}{`survey.design` object}

\item{indicators}{Vector or a List.}

\item{K}{Numerical. Vector of poverty cut-offs. Can be values between 1 and 100.}

\item{weights}{Weights of each indicator.}

\item{over}{Character.}

\item{name}{Character.}

\item{desc}{Character.}

\item{...}{other arguments}
}
\value{
`mpitb_set` object
}
\description{
Set the specification of the project
}
