% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpitb.set.R
\name{mpitb.set}
\alias{mpitb.set}
\title{Set the specification of the project}
\usage{
mpitb.set(
  data,
  indicators,
  weights,
  K = 1,
  subgroup = NULL,
  year = NULL,
  name = "Unnamed project",
  description = "No description",
  ...
)
}
\arguments{
\item{data}{Input data, a `survey.design` class object where a complex survey design was previously specified. Can be a matrix but it is coerced to a `survey.design` class assuming equal probabilities.}

\item{indicators}{A character vector containing the names of the indicators. This character vector should belong to columns names of `data`.}

\item{weights}{A numerical vector representing the weights of each indicator. Should have the same length as `indicators` vector. Should sum up to 1. Default sets the same weight to each indicator}

\item{K}{A numerical vector representing the poverty cut-offs. Should be values between 1 and 100. Default is 1, i.e., deprivations are not censored.}

\item{subgroup}{A character vector containing the names of the populations subgroup of interest. This character vector should belong to columns names of `data`.}

\item{year}{A character object with the name of the column in `data` containing information about temporal information for analysis of changes over time.}

\item{name}{A character object in which it can be specified a name for the project.}

\item{description}{A character object in which it can be specified a brief description for the project.}

\item{...}{other arguments}
}
\value{
`mpitb_set` object
}
\description{
Set the specification of the project
}
\examples{

library(mpitb)
data <- survey::svydesign(id=~PSU, weights = ~Weight, strata = ~Strata,
        data = swz_mics14)
indicators <- c("Water","Assets","School","Nutrition")
weights <- c(1/6,1/6,1/3,1/3)
cutoff <- c(25,50)
subgroup <- c("Region","Area")

set <- mpitb.set(data, indicators, weights, cutoff, subgroup,
      name = "Example", description = "SWZ MICS survey 2014")
}
